Ensure csv/xls mimetypes are present
I was having an issue where the filenames for CSV and XLS files were
being downloaded as the full text of the USE_TITLE_FOR_FILENAME
function. This commit ensures that the xls and csv mime types will be
included regardless of the specific text in the csv/xls download
buttons.